requirements.txt
# Snapchat Memory Fixer Dependencies
# Core dependencies
Pillow>=9.0.0
opencv-python-headless>=4.8.0
piexif>=1.1.3
python-dateutil>=2.8.2

# Optional for video metadata (if needed)
# mutagen>=1.46.0

# For packaging
# pyinstaller>=5.0.0